NyvoNetHunter Project | V : 7.4.5
Features:

    1. Added animation API for the user connection status.

    2. Resolved the API call freezing error ( Issue Note | N : 3 ). 


Code structure:

    1. Replaced the new project back-end API with the old one.
