+++ main_program/source/coding/database/data.py
@@ -1,5 +1,5 @@
 """
-# The base data file of the project
+# The informational module of the project.
 ---
 """
 
@@ -12,38 +12,22 @@ def setup_database_import_path() -> None:
 
     module_paths.append(f"{project_root_directory}/main_program/source/coding")
     module_paths.append(f"{project_root_directory}/main_program/source/coding/database")
-
-
 setup_database_import_path()
 
 
 from exceptions.direct_run_error import DirectRunError
-from packages import (
-    cmd_input,
-    platform,
-)
+from workers.api import clean_terminal
+from packages import dataclass
 
 
-__version__ = "6.2.5"
-__author__ = "KhodeNima ( Nima Bavar )"
-__build_date__ = "2023/11/13"
+@dataclass
+class BuildData:
+    __version__ = "7.4.5"
+    __author__ = "KhodeNima ( Nima Bavar )"
+    __build_date__ = "2023/11/13"
 
-
-def clean_terminal() -> None:
-    runner_operating_system = platform(terse=True)
-    operating_system_is_windows = runner_operating_system.lower().startswith("windows")
-    operating_system_is_linux = runner_operating_system.lower().startswith("linux")
-    
-    if operating_system_is_windows:
-        cmd_input("cls")
-        return
-        
-    if operating_system_is_linux:
-        cmd_input("clear")
-        
 
 module_is_runned_directly = __name__ == "__main__"
-
 if module_is_runned_directly:
     clean_terminal()
     raise DirectRunError(
